Convert GRP files into GMT-formatted strings

GRP files are used by Connectivity Map on-line tool, which stores the information of a rank-ordered list of probesets. They are simply one-column text files, each line containing one probeset. grpFiles2gmt convert GRP files into GMT-formatted strings, which can be written in GMT files to be used by GSEA and other tools.

grp2gmt(txt, chiptype, name)

grpFiles2gmt(..., chiptype, n = -1L)

Arguments

txt

A vector of character strings, each containing one probeset

chiptype

Chip type, normally character representing the microarray chip type. If the option is missing, or is of value NA or NULL, no annotation is done.

name

Character, name of the gene set (the first field of the GMT file)

...

GRP file names

n

Integer, number of lines to be read; -1 indicates all lines should be read

Value

A vector of character strings, each containing one line of a GMT file. They can be written to a file with the writeLines function.

Details

The function grp2gmt, called by grpFiles2gmt internally, annotates probesets when chiptype is supported by GTI, and transform them into the GMT format.

If chiptype is provided, the annotate function is called to fetch probeset annotations from the databank.

Note

It is user's responsibility to check that all GRP files do exist and are readable.

References

See http://www.broadinstitute.org/cmap/index.jsp for the use of GRP files in the Connectivity Map web tool.

Author

Jitao David Zhang <jitao_david.zhang@roche.com>

Examples

up.file <- system.file("extdata/tags_up.grp", package="ribiosExpression")
down.file <- system.file("extdata/tags_down.grp", package="ribiosExpression")
grp2gmt(readLines(up.file, n=-1))
#> [1] "grp\tgrp2gmt\t31730_at\t31731_at\t31732_at\t31733_at"
grpFiles2gmt(c(up.file, down.file), n=3)
#> [1] "tags_up\tgrp2gmt\t31730_at\t31731_at\t31732_at"  
#> [2] "tags_down\tgrp2gmt\t31326_at\t31698_at\t31414_at"

if (FALSE) { # \dontrun{
grp2gmt(readLines(up.file, n=-1), chiptype="HG_U95AV2")
grpFiles2gmt(c(up.file, down.file), n=-1L, chiptype="HG_U95AV2")
} # }