Transform a matrix to long table

Transform a matrix to long table

matrixToLongTable(x, valueLabel = "value", rowLabel = "row", colLabel = "col")

Arguments

x

A matrix

valueLabel

Character string, the label of the value

rowLabel

Character string, the name of the column holding the row names

colLabel

Character string, the name of the column holding the column names

Value

A data.frame

Examples

myMatrix <- matrix(rnorm(24), nrow=4, dimnames=list(LETTERS[1:4], letters[1:6]))
matrixToLongTable(myMatrix)
#>    row col        value
#> 1    A   a -0.100362614
#> 2    B   a  1.374677921
#> 3    C   a  0.359881635
#> 4    D   a -0.911636110
#> 5    A   b  0.372154103
#> 6    B   b  0.824123998
#> 7    C   b -0.277860876
#> 8    D   b  0.290705252
#> 9    A   c -1.462234408
#> 10   B   c -0.936234046
#> 11   C   c -0.005393725
#> 12   D   c  1.146549255
#> 13   A   d -0.300266206
#> 14   B   d -1.672400721
#> 15   C   d  0.342697385
#> 16   D   d  0.004421065
#> 17   A   e -1.329155221
#> 18   B   e -0.755702197
#> 19   C   e -0.401535183
#> 20   D   e -0.142446497
#> 21   A   f -0.614792116
#> 22   B   f -1.294828736
#> 23   C   f -1.256173491
#> 24   D   f  0.480847546