Parse study design and asked questions encoded in design and contrast matrices or in one-way ANOVA designs

Parse study design and asked questions encoded in design and contrast matrices or in one-way ANOVA designs

parseDesignContrast(
  designFile = NULL,
  contrastFile = NULL,
  sampleGroups = NULL,
  groupLevels = NULL,
  dispLevels = NULL,
  contrasts = NULL,
  expSampleNames = NULL
)

Arguments

designFile

A plain tab-delimited file with headers encoding the design matrix, or NULL

contrastFile

A plain tab-delimited file with headers encoding the contrast matrix, or NULL

sampleGroups

A character string concatenated by commas (e.g. A,B,C), or a plain text file containing one string per line (e.g. AnewlineBnewlineC), encoding sample group memberships.

groupLevels

Similar format as 'sampleGroups', encoding levels (e.g. order) of the sampleGroups

dispLevels

Similar format as 'sampleGroups', encoding the display of the groupLevels. Must match 'groupLevels'

contrasts

Similar format as 'sampleGroups', encoding contrasts in case of one-way ANOVA designs

expSampleNames

A vector of character strings giving the expected sample names (e.g. those in the input matrix)

Value

A S4-object 'DesignContrast'

Examples

## one-way ANOVA
parseDesignContrast(sampleGroups="As,Be,As,Be,As,Be",groupLevels="Be,As",
    dispLevels="Beryllium,Arsenic", contrasts="As-Be")
#> DesignContrast object:
#> - 6 samples in 2 groups
#>     Levels: Be, As
#> - Design matrix (6 samples x 2 variables)
#>     Variables: Be, As
#>   Call 'designMatrix(object)' to get the design matrix.
#> - Contrast matrix (2 variables x 1 contrasts)
#>     Contrasts: As-Be
#>   Call 'contrastMatrix(object)' to get the contrast matrix.
#>   Call 'contrastAnnotation(object) to get the contrast annotation.
## design/contrast matrix
designFile <- system.file("extdata/example-designMatrix.txt",
    package="ribiosExpression")
contrastFile <- system.file("extdata/example-contrastMatrix.txt",
    package="ribiosExpression")
# minimal information
parseDesignContrast(designFile=designFile, contrastFile=contrastFile)
#> DesignContrast object:
#> - 6 samples in 2 groups
#>     Levels: AutoGroup_01, AutoGroup_02
#> - Design matrix (6 samples x 2 variables)
#>     Variables: Be, As
#>   Call 'designMatrix(object)' to get the design matrix.
#> - Contrast matrix (2 variables x 1 contrasts)
#>     Contrasts: As.vs.Be
#>   Call 'contrastMatrix(object)' to get the contrast matrix.
#>   Call 'contrastAnnotation(object) to get the contrast annotation.
# with extra information about sample groups
parseDesignContrast(designFile=designFile, contrastFile=contrastFile,
    sampleGroups="As,Be,As,Be,As,Be",
    groupLevels="Be,As", dispLevels="Beryllium,Arsenic")
#> DesignContrast object:
#> - 6 samples in 2 groups
#>     Levels: Be, As
#> - Design matrix (6 samples x 2 variables)
#>     Variables: Be, As
#>   Call 'designMatrix(object)' to get the design matrix.
#> - Contrast matrix (2 variables x 1 contrasts)
#>     Contrasts: As.vs.Be
#>   Call 'contrastMatrix(object)' to get the contrast matrix.
#>   Call 'contrastAnnotation(object) to get the contrast annotation.