Read table with Foreign Keys

The concept Foreign Keys comes from relational databse systems. These keys can be used to cross-reference tables. Say we have two data.frames, one contains gene annotations and the other contains protein annotations. A column named mRNArefseqID may be the foreign key that can be used to specify relationships between gene and proteins.

readFKtable(file, fk, strict.order = FALSE, ...)

Arguments

file

A table file.

fk

Characters, foreign keys.

strict.order

Logical, whether the foreign keys must have the same order as they appear in the file.

...

Other parameters passed to the read.table function.

Value

A data.frame if the FK-matching was successful, otherwise the function will print an error message and stop.

Details

The readFKtable reads a table from file, and checks if it contains provided foreign keys: either as row.names or in the first column.

Author

Jitao David Zhang <jitao_david.zhang@roche.com>

Examples

test.file <- tempfile()
fk.teams <- c("HSV", "FCB", "BVB")
## FK in row names
test.mat <- matrix(rnorm(9), nrow=3, dimnames=list(fk.teams, NULL))
write.table(test.mat, test.file)
readFKtable(test.file, fk=fk.teams)
#>             V1         V2           V3
#> HSV  0.3598816  0.8241240 -1.462234408
#> FCB -0.9116361 -0.2778609 -0.936234046
#> BVB  0.3721541  0.2907053 -0.005393725

## or: FK can be in the first column
test.df <- data.frame(team=fk.teams, pts=c(15,14,15),plc=c("H", "G", "H"))
write.table(test.df, test.file)
readFKtable(test.file, fk=fk.teams)
#>     pts plc
#> HSV  15   H
#> FCB  14   G
#> BVB  15   H

## try strict.order=TRUE
test.df <- data.frame(pts=c(15,14,13), plc=c("H", "G", "H"), row.names=rev(fk.teams))
write.table(test.df, test.file)
readFKtable(test.file, fk=fk.teams, strict.order=FALSE)
#>     pts plc
#> HSV  13   H
#> FCB  14   G
#> BVB  15   H
if (FALSE) readFKtable(test.file, fk=fk.teams, strict.order=TRUE) # \dontrun{}