Export ExpressionSet as Gct/Cls files

Gct/Cls file formats are required by the Gene Set Enrichment Analysis (GSEA) tool. Functions writeGct and writeCls exports file of two formats respectively, and writeGctCls calls the two function internally to write two files.

writeCls(eset, file = stdout(), sample.group.col = "group")

writeGctCls(
  eset,
  file.base,
  feat.name,
  feat.desc,
  sample.group.col,
  write.add.fData.file = TRUE,
  write.add.pData.file = TRUE
)

Arguments

eset

An object of the eSet class, for example an ExpressionSet object

file

Name of the Gct/Cls file. If left missing, the file is printed on the standard output.

sample.group.col

Integer, character or a factor vector of the same length as the sample number, indicating classes (groups) of samples. See details.

file.base

For writeGctCls, the base name of the two files: the suffix (.gct and .cls) will be appended

feat.name

Integer or character, indicating which column of the featureData should be used as feature descriptions. If the value is missing, the Description column of the Gct file will be left blank

feat.desc

Integer or character, indicating which column of the featureData should be used as feature names; if missing, results of the featureNames function will be used as identifiers in the Gct file. See details.

write.add.fData.file

Logical, whether additional featureData should be written into a file named ${file.base}.add.fData.txt

write.add.pData.file

Logical, whether additional phenoData should be written into a file named ${file.base}.add.pData.txt

Value

Functions are used for their side effects.

Details

The feat.name option specifies what identifiers should be used for features (probesets). When the value is missing, featureNames is called to provide feature identifiers.

In contrast, the sample.group.col cannot be missing: since cls files encode groups (classes) of samples, and if sample.group.col was missing, it is usually impossible to get class information from sampleNames.

Internally writeCls calls dfFactor function to determine factor of samples. Therefore sample.group.col is to a certain degree generic: it can be a character string or integer index of the pData(eset) data matrix, or a factor vector of the same length as ncol(eset).

Functions

writeCls(): writeCls

References

http://www.broadinstitute.org/gsea/doc/GSEAUserGuideTEXT.htm

See also

Author

Jitao David Zhang <jitao_david.zhang@roche.com>

Examples

data(sample.ExpressionSet, package="Biobase")
writeGct(sample.ExpressionSet[1:5, 1:4], file=stdout())
#> #1.2
#> 5	4
#> NAME	Description	A	B	C	D
#> AFFX-MurIL2_at		192.742	85.7533	176.757	135.575
#> AFFX-MurIL10_at		97.137	126.196	77.9216	93.3713
#> AFFX-MurIL4_at		45.8192	8.83135	33.0632	28.7072
#> AFFX-MurFAS_at		22.5445	3.60093	14.6883	12.3397
#> AFFX-BioB-5_at		96.7875	30.438	46.1271	70.9319
writeCls(sample.ExpressionSet, file=stdout(), sample.group.col="type")
#> 26 2 1
#> # Case Control
#> 1 0 1 0 0 1 0 0 0 1 0 1 0 0 0 1 0 1 0 0 1 1 1 1 0 0

tmpfile <- tempfile()
writeGctCls(sample.ExpressionSet, file.base=tmpfile, sample.group.col="type")
readLines(paste(tmpfile, ".cls",sep=""))
#> [1] "26 2 1"                                             
#> [2] "# Case Control"                                     
#> [3] "1 0 1 0 0 1 0 0 0 1 0 1 0 0 0 1 0 1 0 0 1 1 1 1 0 0"
unlink(c(paste(tmpfile, ".cls", sep=""), paste(tmpfile, ".gct", sep="")))